+++ Package.swift
@@ -7,9 +7,7 @@ let package = Package(
     name: "XPlatform",
     platforms: [
         .iOS(.v15),
-        .macOS(.v12),
-        .tvOS(.v15),
-        .watchOS(.v8)
+        .macOS(.v12)
     ],
     products: [
         // Products define the executables and libraries a package produces, making them visible to other packages.