+++ Sources/XPlatform/XPasteboard+Extensions.swift
@@ -14,17 +14,9 @@ import UIKit
 #endif
 
 extension XPasteboard {
-	/// Returns the general/system pasteboard
-	public static var xGeneral: XPasteboard {
-		#if canImport(AppKit)
-		return NSPasteboard.general
-		#else
-		return UIPasteboard.general
-		#endif
-	}
-	
-	/// Gets or sets the string content of the pasteboard
-	public var xString: String? {
+	/// Gets or sets the string content of the pasteboard.
+	/// Unified name across `NSPasteboard` and `UIPasteboard`.
+	public var stringValue: String? {
 		get {
 			#if canImport(AppKit)
 			return self.string(forType: .string)